+++ app.go
@@ -5,13 +5,19 @@ import (
 	"fmt"
 	"net/http"
 	"strings"
+
+	"github.com/bchadwic/wordbubble/auth"
+	"github.com/bchadwic/wordbubble/model"
+	"github.com/bchadwic/wordbubble/user"
+	"github.com/bchadwic/wordbubble/util"
+	"github.com/bchadwic/wordbubble/wb"
 )
 
 type App struct {
-	logger Logger
-	auth   Auth
-	users  Users
-	wbs    WordBubbles
+	logger util.Logger
+	auth   auth.Auth
+	users  user.Users
+	wbs    wb.WordBubbles
 }
 
 func (app *App) respond(response string, statusCode int, w http.ResponseWriter) {
@@ -26,14 +32,14 @@ func (app *App) Signup(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var user User
+	var user model.User
 	if err := json.NewDecoder(r.Body).Decode(&user); err != nil {
 		app.logger.Error("could not decode user from body: %s", err)
 		app.respond("could not parse a user from the request body", http.StatusBadRequest, w)
 		return
 	}
 
-	if err := ValidUser(&user); err != nil {
+	if err := util.ValidUser(&user); err != nil {
 		app.respond(err.Error(), http.StatusBadRequest, w)
 		return
 	}
@@ -136,9 +142,9 @@ func (app *App) Token(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	fmt.Printf("timeBeforeExpiration: %d, ImminentExpirationWindow: %d\n", timeBeforeExpiration, ImminentExpirationWindow)
+	fmt.Printf("timeBeforeExpiration: %d, ImminentExpirationWindow: %d\n", timeBeforeExpiration, auth.ImminentExpirationWindow)
 	var latestRefreshToken string
-	if timeBeforeExpiration < ImminentExpirationWindow {
+	if timeBeforeExpiration < auth.ImminentExpirationWindow {
 		latestRefreshToken = app.auth.GetOrCreateLatestRefreshToken(userId)
 	}
 
@@ -185,7 +191,7 @@ func (app *App) Push(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var wb WordBubble // finally we are authenticated! Let's insert a wordbubble
+	var wb model.WordBubble // finally we are authenticated! Let's insert a wordbubble
 	if err = json.NewDecoder(r.Body).Decode(&wb); err != nil {
 		app.logger.Error("could not decode wordbubble from body: %s", err)
 		app.respond("could not parse a wordbubble from request body", http.StatusBadRequest, w)
