+++ user/data.go
@@ -1,1 +1,2 @@
 package user
+
